requirements.txt
flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
numpy>=1.23
Pillow>=9.0.0
tensorflow>=2.12.0 